Refactor code for getting manifest/result content size
